tried moving lgoger and prisma to lib, but defining import in nuxt config doesnt work
