+++ server/plugins/logger-init.ts
@@ -1,4 +1,4 @@
-import { logger } from "../utils/loggers"
+import { logger } from "../lib/loggers"
 
 export default defineNitroPlugin(() => {
   logger.info("Logger initialized")
