+++ server/seeds/user.ts
@@ -1,5 +1,6 @@
 import { v4 as uuidv4 } from "uuid"
 import Seed from "./seed"
+import prisma from "../lib/prisma"
 
 /**
  * User seed.
