+++ server/utils/auth.ts
@@ -3,7 +3,7 @@ import { betterAuth } from "better-auth"
 import { APIError } from "better-auth/api"
 import { admin } from "better-auth/plugins"
 import { prismaAdapter } from "better-auth/adapters/prisma"
-import prisma from "../../server/utils/prisma"
+import prisma from "../lib/prisma"
 
 /**
  * betterAuth config
